requirements.txt
Django==1.11.22 # pyup: >=1.11,<1.12
django_extensions==2.2.1
django-widget-tweaks==1.4.5
icalendar==4.0.3
psycopg2-binary==2.8.3
pytz==2019.1
lockfile==0.12.2
django-environ==0.4.5
